refactor(viewer): apply viewer style to Measurement Viewer Drawer
